+++ config/http/errors.php
@@ -6,4 +6,4 @@ use Opulence\Debug\Errors\Handlers\ErrorHandler;
  * Define the error handler
  * ----------------------------------------------------------
  */
-return new ErrorHandler($exceptionHandler);
+return new ErrorHandler($logger, $exceptionHandler);